record write read

+++ internal/store/store_operator.go
@@ -3,7 +3,8 @@ package store
 import "bytes"
 
 type store_operator interface {
-	WriteData(buf *bytes.Buffer) (int, error)
+
+	WriteData(buf *bytes.Buffer) (int64, error)
 
     ReadData(buf *bytes.Buffer) (interface{}, error)
 }
